Frame writers: use getObject when reading string/complex types.
This mirrors similar logic in numeric aggregators. (The same method
from AggregatorUtil is even used to determine when to apply the logic.)

The idea is that when an underlying selector is STRING or COMPLEX typed,
we should call getObject and cast the result to number, rather than using
the primitive numeric accessor methods.

This fixes an issue where a column would be read as all-zeroes when its
SQL type is numeric, and its physical type is string. This can happen when
evolving a column's type from string to number.