Add support to optionally parse numbers from strings for text-based parsers.
This helps samplers to detect numeric types for text-based formats like
csv and tsv.

These text-based formats by default parse numbers as strings.
This change add a config flag to optionally parse numbers as numbers.
Long for integers and Double for floating-point numbers.
It falls back to string if it cannot parse.

The web-console has some code in the load data flow to parse the sample of data
returned by the Druid sampler to further inspect types so it can convert them
to specific numeric types, if applicable.

After this change, the web-console sampler/other applications can just rely
on Druid to do it.