projections part 1 - realtime segments
This PR begins to introduce the concept of projections to Druid datasources, which are similar to materialized views but are built into a segment, and which can automatically be used during query execution if the projection fits the query. This PR only contains the logic to build and query them for realtime queries, and does not contain the ability to serialize and actually store them in persisted segments, so it is effectively a toy right now.

changes:
* Adds ProjectionSpec interface, AggregateProjectionSpec implementation for defining rollup projections on druid datasources
* Adds projections to DataSchema
* Adds projection building and querying support to OnHeapIncrementalIndex