Web console: Fixed sampling for delta source in classic data loader and MSQ (#17160)
Added the delta tile to MSQ and made sure it works there and classic data loader. It is special because it always assumes a parquet format so it has to skip some steps in the code. Now the framework is there to add more tiles that force a specific format