MSQ: Add exception messages to WorkerFailedFault and WorkerRpcFailedFault.
It's useful for the fault message to have some description of the
underlying error: why did the worker fail, or why did the RPC call fail?
This patch updates the fault messages to include the toString of the error.

The full stack trace for these faults is available to end users in
"exceptionStackTrace" of the report. But for cluster operators, if debug
logging is not enabled, then for Dart only the message is logged, not the
stack trace. In this scenario it's useful for the message to be more
detailed.