KubernetesTaskRunner: Wait in start() for tasks to be located.
This helps with orderly Overlord failover. After this patch, by the time
the task runner returns from start(), all tasks are located (subject to some
timeout). This is useful for supervisors, which start next and which need
to contact tasks.

The timeout defaults to 1 minute, and is configurable using
druid.indexer.runner.taskJoinTimeout.