Remove cronExecutor to keep things simple.
The jobsExecutor is reponsible for submitting jobs to the broker for all
scheduled batch supervisors. The cronExecutor was simply decoupling the
submitting of jobs from the actual scheduled running of jobs. The service client
library already has async threads, so we remove the cronExecutor to keep things
simple as things are handled in an async manner by the service client. If we ever
observe and see evidence of  bottlenecks around task submission, etc, we should be
able to make jobsExecutor multiple threaded instead of single threaded.