fix issue with OR filter vector value matcher for proper 3VL behavior when inverted (#17655) (#17656)
Fixes a bug with the OR filters vectorized value matcher that causes vector engines processing filter an OR filter under a NOT filter (ex. of the form NOT(x OR y)) to produce incorrect results for null values matched.

This bug is due to incorrectly hard coding the includeUnknown parameter as false for OR filter child vector matchers after the initial filter clause instead of passing it through the function parameter to the underlying matchers.