Add druid.coordinator.kill.maxInterval property.
This property allows interval-based limiting of the number of segments
killed, as well as the already-existing segment-count-based limiting.
This is useful because the kill task retrieves payloads of *all* active
segments that fall within the kill interval, which can be a large number.

The new default is 30 days, which seems reasonable for common situations.