Various fixes for large columns.
This patch fixes a class of bugs where various primitive column readers were
not providing a SmooshedFileMapper to GenericIndexed, even though the corresponding
writer could potentially write multi-file columns. For example, #7943 is an
instance of this bug.

This patch also includes a fix for an issue on the writer for compressed
multi-value string columns, V3CompressedVSizeColumnarMultiIntsSerializer, where it
would use the same base filename for both the offset and values sections. This bug
would only be triggered for segments in excess of 500 million rows. When a segment
has fewer rows than that, it could potentially have a values section that needs
to be split over multiple files, but the offset is never more than 4 bytes per row.
This bug was triggered by the new tests, which use a smaller fileSizeLimit.