DictionaryEncodedColumnPartSerde: Read values using smooshReader.
The write side uses a FileSmoosher for valueWriter, which means that the
value component may be split over multiple files. In this case, the
read side needs a SmooshedFileMapper.

It is plausible for multi-valued columns to have value sections in excess
of 2 GB: imagine a segment with 10 million rows and an average of 50 values
per row. The value section would have 500 million ints, which is 2 GB.