Add check to see if druid is running in kubernetes (#17697)
The PR #16386 changed the docker file to use the canonical hostname instead of ip by default. This was an attempt to fix druid.indexer.task.restoreTasksOnRestart , as it does not work on kubernetes without the canonical hostname.

This change is not backward compatible, as switching to canonical names would not work with docker. This PR attempts to fix this by adding a check to see if the deployment is in a kubernetes environment by checking ${KUBERNETES_SERVICE_HOST} and using canonical hostnames instead of IP addresses only in that case.